+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+# v2.3.0
+
 * Allow using inline render method(s) defined on a parent.
 
     *Simon Rand*